Hardcode all keys as released
This is a temporary fix, since 0 is key pressed and for some reason the
games won’t load is all keys are pressed.

Now the Zelda LTTP and Pokemon Emerald both load, but Pokemon gets
stuck at the GameFreak logo. The graphics are also a bit messed up, but
I’m fixing them next.