Fix a problem with background layer transforms
Basically, the hardware doesn’t actually apply it as a matrix
multiplied by a vector, it actually uses increments on each line and
pixel. This is a problem when the reference point registers are changed
mid frame, because it effectively resets the incrementation, which is
used by game like mario cart.